src/NlwExpertAuction.API/UseCases/Offers/CreateOffer/CreateOfferUseCase.cs
using NlwExpertAuction.API.Comunication.Requests;
using NlwExpertAuction.API.Contracts;
using NlwExpertAuction.API.Entities;
using NlwExpertAuction.API.Services;

namespace NlwExpertAuction.API.UseCases.Offers.CreateOffer;

public class CreateOfferUseCase
{
    private readonly ILoggedUser _loggedUser;
    private readonly IOfferRepository _repository;

    public CreateOfferUseCase(ILoggedUser loggedUser, IOfferRepository repository)
    {
        _loggedUser = loggedUser;
        _repository = repository;
    }

    public int Execute(int itemId, RequestCreateOfferJson request)
    {
        var user = _loggedUser.User();

        var offer = new Offer
        {
            CreatedOn = DateTime.Now,
            ItemId = itemId,
            Price = request.Price,
            UserId = user.Id
        };

        _repository.Add(offer);

        return offer.Id;
    }
}
